Players could hit their ships right away when firing the Phoenix or guided missiles in slow motion or bullet time modes. Tweaked shot coronas and light trails some more.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@1189 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/objrender.c
@@ -598,7 +598,7 @@ int DrawPolygonObject (tObject *objP, int bDepthSort)
 
 if (objP->nType == 255)
 	return 0;
-if (0 > (id = objP->id))
+if (0 > (id = (int) objP->id))
 	bEnergyWeapon = id = 0;
 else 
 	bEnergyWeapon = (objP->nType == OBJ_WEAPON) && gameData.objs.bIsWeapon [id] && !gameData.objs.bIsMissile [id];
